requirements.txt
requests 
beautifulsoup4
lxml
langchain==0.3.26
langchain-community==0.3.21
langchain-core==0.3.68
langchain-google-genai==2.1.7
langchain-mcp-adapters==0.1.8
langchain-ollama==0.3.4
langchain-postgres==0.0.15
langgraph==0.5.2
python-dotenv
psycopg[binary]==3.2.9
psycopg2-binary==2.9.10
fastmcp==2.10.1
pydantic==2.11.7
fastapi
uvicorn
langchain-mcp-adapters==0.1.8
rich
langgraph-checkpoint-postgres==2.0.21
supabase
openai
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
uuid